+++ src/test/java/org/opensearch/knn/index/query/iterators/NestedBinaryVectorIdsKNNIteratorTests.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.knn.index.query.filtered;
+package org.opensearch.knn.index.query.iterators;
 
 import junit.framework.TestCase;
 import lombok.SneakyThrows;
@@ -17,10 +17,13 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class NestedFilteredIdsKNNByteIteratorTests extends TestCase {
+public class NestedBinaryVectorIdsKNNIteratorTests extends TestCase {
     @SneakyThrows
     public void testNextDoc_whenIterate_ReturnBestChildDocsPerParent() {
         final SpaceType spaceType = SpaceType.HAMMING;
@@ -45,7 +48,7 @@ public class NestedFilteredIdsKNNByteIteratorTests extends TestCase {
         }
 
         // Execute and verify
-        NestedFilteredIdsKNNByteIterator iterator = new NestedFilteredIdsKNNByteIterator(
+        NestedBinaryVectorIdsKNNIterator iterator = new NestedBinaryVectorIdsKNNIterator(
             filterBitSet,
             queryVector,
             values,
@@ -57,5 +60,32 @@ public class NestedFilteredIdsKNNByteIteratorTests extends TestCase {
         assertEquals(filterIds[2], iterator.nextDoc());
         assertEquals(expectedScores.get(2), iterator.score());
         assertEquals(DocIdSetIterator.NO_MORE_DOCS, iterator.nextDoc());
+    }
+
+    @SneakyThrows
+    public void testNextDoc_whenIterateWithoutFilters_thenReturnBestChildDocsPerParent() {
+        final SpaceType spaceType = SpaceType.HAMMING;
+        final byte[] queryVector = { 1, 2, 3 };
+        // Parent id for 0 -> 1
+        // Parent id for 2, 3 -> 4
+        // In bit representation, it is 10010. In long, it is 18.
+        final BitSet parentBitSet = new FixedBitSet(new long[] { 18 }, 5);
+        final List<byte[]> dataVectors = Arrays.asList(new byte[] { 11, 12, 13 }, new byte[] { 14, 15, 16 }, new byte[] { 17, 18, 19 });
+        final List<Float> expectedScores = dataVectors.stream()
+            .map(vector -> spaceType.getKnnVectorSimilarityFunction().compare(queryVector, vector))
+            .collect(Collectors.toList());
+
+        KNNBinaryVectorValues values = mock(KNNBinaryVectorValues.class);
+        when(values.getVector()).thenReturn(dataVectors.get(0), dataVectors.get(1), dataVectors.get(2));
+        when(values.nextDoc()).thenReturn(0, 2, 3, Integer.MAX_VALUE);
+
+        // Execute and verify
+        NestedBinaryVectorIdsKNNIterator iterator = new NestedBinaryVectorIdsKNNIterator(queryVector, values, spaceType, parentBitSet);
+        assertEquals(0, iterator.nextDoc());
+        assertEquals(expectedScores.get(0), iterator.score());
+        assertEquals(3, iterator.nextDoc());
+        assertEquals(expectedScores.get(2), iterator.score());
+        assertEquals(DocIdSetIterator.NO_MORE_DOCS, iterator.nextDoc());
+        verify(values, never()).advance(anyInt());
     }
 }